We changed from getStaticProps to getServerSideProps and added a helper to check for the cookie
